Move services to new services/ dir
