+++ streaming/__init__.py
@@ -5,10 +5,10 @@ from .streaming import (  # noqa: F401
     StreamingServiceTrack,
 )
 
-from .service_gmusic import GMusic, GMusicTrack  # noqa: F401
-from .service_youtube import YouTube, YouTubeTrack  # noqa: F401
-from .service_ytmusic import YTMusic, YTMusicTrack  # noqa: F401
-from .service_spotify import Spotify, SpotifyTrack  # noqa: F401
+from .service.gmusic import GMusic, GMusicTrack  # noqa: F401
+from .service.youtube import YouTube, YouTubeTrack  # noqa: F401
+from .service.ytmusic import YTMusic, YTMusicTrack  # noqa: F401
+from .service.spotify import Spotify, SpotifyTrack  # noqa: F401
 
 SUPPORTED_STREAMING_SERVICES = StreamingService.__subclasses__()
